Colocate types
