--- src/@types/MemoryFiles.d.ts
@@ -1,21 +1,0 @@
-import type { Snowflake } from "discord.js";
-
-/**
- * These types define the format of the JSON memory files that should be stored in the account directory for the bot.
- */
-
-/**
- * The format of the config JSON file.
- * See an example in [accounts/ExampleUsername/config.json](../../accounts/ExampleUsername/config.json).
- */
-export interface ConfigFile {
-	clientId: Snowflake;
-	token: string;
-	url: string;
-}
-
-/**
- * The format of the whitelist JSON file.
- * See an example in [accounts/ExampleUsername/whitelist.json](../../accounts/ExampleUsername/whitelist.json).
- */
-export type WhitelistFile = Snowflake[];
