+++ src/commands/CommandHandler.ts
@@ -1,23 +1,21 @@
-import { SlashCommandBuilder } from "discord.js";
+import { SlashCommandBuilder, type Awaitable } from "discord.js";
 import type { ChatInputCommandInteraction } from "discord.js";
 
 import { replyWithError } from "../helpers/replyWithError.js";
 import { error } from "../logger.js";
 
-type CommandExecution = (interaction: ChatInputCommandInteraction) => Promise<void>;
-
 export class CommandHandler extends SlashCommandBuilder {
-	private execution?: CommandExecution;
+	private execution?: typeof this.execute;
 
-	public setExecution (execution: CommandExecution): this {
+	public setExecution (execution: typeof this.execute): this {
 		this.execution = execution;
 		return this;
 	}
 
-	public async execute (interaction: ChatInputCommandInteraction): Promise<void> {
+	public execute (interaction: ChatInputCommandInteraction): Awaitable<void> {
 		if (this.execution) {
 			try {
-				await this.execution(interaction);
+				return this.execution(interaction);
 			}
 			catch (err) {
 				error(err);
