+++ src/commands/help.ts
@@ -1,7 +1,7 @@
 import { EmbedBuilder, userMention } from "discord.js";
 import type { User } from "discord.js";
 
-import { CommandHandler } from "../@types/CommandHandler.js";
+import { CommandHandler } from "./CommandHandler.js";
 import { lastUpdated, embedColors, version } from "../parameters.js";
 
 export const help = new CommandHandler()
