+++ src/commands/unwhitelist.ts
@@ -1,6 +1,6 @@
 import { EmbedBuilder, channelMention } from "discord.js";
 
-import { CommandHandler } from "../@types/CommandHandler.js";
+import { CommandHandler } from "./CommandHandler.js";
 import { deleteContext } from "../memory/context.js";
 import { removeChannel as unwhitelistChannel } from "../memory/whitelist.js";
 import { embedColors } from "../parameters.js";
