+++ src/commands/whitelist.ts
@@ -1,6 +1,6 @@
 import { EmbedBuilder, channelMention } from "discord.js";
 
-import { CommandHandler } from "../@types/CommandHandler.js";
+import { CommandHandler } from "./CommandHandler.js";
 import { generateContext } from "../memory/context.js";
 import { addChannel as whitelistChannel } from "../memory/whitelist.js";
 import { embedColors } from "../parameters.js";
